+++ backend/src/main/java/org/example/repository/nosql/MessageRepository.java
@@ -13,7 +13,7 @@ public interface MessageRepository extends MongoRepository<Message, String> {
             "{ 'senderId': ?0, 'recipientId': ?1 }, " +
             "{ 'senderId': ?1, 'recipientId': ?0 } " +
             "] }")
-    List<Message> findChatHistory(Integer u1, Integer u2);
+    List<Message> findChatHistory(Long u1, Long u2);
 
-    Message findFirstBySenderIdAndRecipientIdOrderBySentAtDesc(Integer senderId, Integer recipientId);
+    Message findFirstBySenderIdAndRecipientIdOrderBySentAtDesc(Long senderId, Long recipientId);
 }
